chore: started cleaning
