+++ services/frontend/app/package.json
@@ -5,12 +5,14 @@
   "description": "",
   "main": "server.js",
   "scripts": {
+    "css": "npx @tailwindcss/cli -i ./src/css/src.css -o ./src/css/output.css",
     "dev": "nodemon"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@fastify/static": "^8.2.0",
+    "@tailwindcss/cli": "^4.1.13",
     "@types/node": "^24.3.0",
     "fastify": "^5.5.0",
     "tailwindcss": "^4.1.12",
