+++ services/frontend/app/src/client/scripts/404.ts
@@ -1,7 +1,7 @@
 export function render404(main: HTMLElement) {
   main.innerHTML = `
   <div class="container">
-    <img src="/images/capybara.webp" alt="capy" style="width:100%;">
+    <img src="/capybara.webp" alt="capy" style="width:100%;">
   </div>
   `
 }
