--- services/frontend/app/src/client/scripts/central.ts
@@ -1,24 +1,0 @@
-
-import { renderPageTemplate } from './page.template.ts'
-import { t } from './language/translation.ts'
-import { renderLanguageDropdownButton } from './language/languageDropdownButton.ts'
-
-export function renderCentral(main: HTMLElement) {
-  renderPageTemplate(main, {
-	title: t('CENTRAL'),
-	nextButtons: [
-	  { href: "/game/tournament", label: t('Tournament') },
-	  { href: "/quickMatch", label: t('Quick Match') },
-	  { href: "/account", label: t('Profile') },
-	  { href: "/game/leaderboard", label: t('Leaderboard') },
-	  { href: "/game/match", label: t('Game') },
-	],
-	backHref: "/",
-	showBack: true,
-	homeHref: "/"
-  });
-  	const langDropdown = document.getElementById('lang-dropdown-container');
-  if (langDropdown) {
-	renderLanguageDropdownButton(langDropdown, () => renderCentral(main));
-  }
-}
