+++ services/frontend/app/src/client/scripts/language/translation.ts
@@ -3,9 +3,9 @@ type TranlationMap = Record<string, string>;
 let translation: TranlationMap = {};
 export let currentLang: string = "en";
 
-import en from '../../assets/locales/en.json' with { type: "json" };
-import fr from '../../assets/locales/fr.json' with { type: "json" };
-import es from '../../assets/locales/es.json' with { type: "json" };
+import en from '../../../assets/locales/en.json' with { type: "json" };
+import fr from '../../../assets/locales/fr.json' with { type: "json" };
+import es from '../../../assets/locales/es.json' with { type: "json" };
 
 const resources: Record<string, TranlationMap> = { en ,fr, es };
 
@@ -16,7 +16,7 @@ const savedLang = localStorage.getItem('selectedLanguage');
   } else {
     currentLang = 'en';
   }
-  translation = resources[currentLang];
+  translation = resources[currentLang] ?? {};
 }
 
 export function t(key: string): string {
