--- services/frontend/app/src/client/scripts/leaderboard.ts
@@ -1,13 +1,0 @@
-
-import { renderPageTemplate } from './page.template.ts'
-
-export function renderLeaderboard(main: HTMLElement) {
-	renderPageTemplate(main, {
-		title: "LEADERBOARD",
-		nextButtons: [
-		],
-		backHref: "/central",
-		showBack: true,
-		homeHref: "/"
-	});
-}
