--- services/frontend/app/src/client/scripts/tournament.ts
@@ -1,17 +1,0 @@
-
-import { renderPageTemplate } from './page.template.ts'
-
-export function renderTournament(main: HTMLElement) {
-	renderPageTemplate(main, {
-		title: "TOURNAMENT",
-		nextButtons: [
-			{ href: "/404", label: "4 players" },
-			{ href: "/404", label: "8 players" },
-			{ href: "/404", label: "16 players" },
-
-		],
-		backHref: "/central",
-		showBack: true,
-		homeHref: "/"
-	});
-}
