+++ services/frontend/app/src/pages/404HTML.ts
@@ -1,7 +1,7 @@
 export function render404(): string {
   return `
   <div class="container">
-    <img src="/images/capybara.webp" alt="capy" style="width:100%;">
+    <img src="/capybara.webp" alt="capy" style="width:100%;">
   </div>
   `
 }
