--- services/frontend/app/src/pages/leaderboard.ts
@@ -1,13 +1,0 @@
-
-import { renderPageTemplate } from './page.template.js'
-
-export function renderLeaderboard(): string {
-	return renderPageTemplate({
-		title: "LEADERBOARD",
-		nextButtons: [
-		],
-		backHref: "/central",
-		showBack: true,
-		homeHref: "/"
-	});
-}
