--- services/frontend/app/src/pages/profile.ts
@@ -1,16 +1,0 @@
-
-import { renderPageTemplate } from './page.template.js'
-import { translate } from './language/translation.js'
-
-export function renderProfile(): string {
-	return renderPageTemplate({
-		title: "PROFILE",
-		nextButtons: [
-			{ href: "/accountSettings", label: translate("Account Settings") },
-			{ href: "/editProfile", label: translate("Edit Profile") }
-		],
-		backHref: "/central",
-		showBack: true,
-		homeHref: "/"
-	});
-}
