--- services/frontend/app/src/pages/tournament.ts
@@ -1,17 +1,0 @@
-
-import { renderPageTemplate } from './page.template.js'
-
-export function renderTournament(): string {
-	return renderPageTemplate({
-		title: "TOURNAMENT",
-		nextButtons: [
-			{ href: "/404", label: "4 players" },
-			{ href: "/404", label: "8 players" },
-			{ href: "/404", label: "16 players" },
-
-		],
-		backHref: "/central",
-		showBack: true,
-		homeHref: "/"
-	});
-}
