+++ services/frontend/app/src/server/server.ts
@@ -2,7 +2,7 @@ import Fastify from 'fastify'
 import type { FastifyInstance } from 'fastify';
 import { servRoutes } from './route.js';
 import { options } from './serv.conf.js';
-// import fastifyStatic from '@fastify/static'
+import fastifyStatic from '@fastify/static'
 import fastifyVite from '@fastify/vite'
 
 try {
@@ -21,11 +21,16 @@ catch (err) {
 
 async function addPlugins(serv: FastifyInstance) {
 	await serv.register(servRoutes)
-				// .register(fastifyStatic, {
-				// 	root: [
-				// 		'/app/src/client/pages'
-				// 	]
-				// })
+				.register(fastifyStatic, {
+					root: [
+						"/app/src",
+						"/app/src/css",
+						"/app/dist/client/spaRouter",
+						"/app/dist/client/scripts",
+						"/app/dist/pages",
+						"/app/src/images"
+					]
+				})
 				// .register(fastifyVite, {
 				// 	root: "/app",
 				// 	dev: process.argv.includes('--dev'),
