+++ services/frontend/app/tailwind.config.js
@@ -1,0 +1,10 @@
+// tailwind.config.js
+export default {
+  content: [
+    "./pages/**/*.{js,ts,jsx,tsx}",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
