+++ services/frontend/tools/front.sh
@@ -8,5 +8,6 @@ npm install
 
 # npx tsc
 
+npm run css
 npm run dev
 #npm start
