+++ services/nginx/conf/default.conf
@@ -57,7 +57,7 @@ server {
 		
 		# forbids browser from guessing the content type of a payload, limiting 
 		# innocent or malicious misinterpretations.
-		add_header X-Content-Type-Options "nosniff" always;
+		# add_header X-Content-Type-Options "nosniff" always;
 
 		# defines that the only origin allowed to load the requested ressource
 		# is on the same host, protocol and port as the server.
